Separate common classes for server and notebook extensions
